requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=8.0
pytest-cov>=5.0
python-dotenv>=1.0